x64 Release Build

+++ TSS.MSR/TSS.CPP.Server/Samples/utilities.h
@@ -32,6 +32,7 @@ typedef unsigned char BYTE;
 #include <cassert>
 #include <fstream>
 #include <limits.h>
+#include <chrono>
 
 
 
@@ -195,7 +196,7 @@ std::vector<std::vector<BYTE> > splitChunks(const std::vector<BYTE> & v, unsigne
 		x.reserve(chunksize);
 	}
 	for (unsigned int i = 0; i < v.size(); i += chunksize) {
-		unsigned int endpoint = std::min(v.size(), i + chunksize);
+		unsigned int endpoint = std::min((unsigned int)v.size(), i + chunksize);
 		ret[i / chunksize].insert(ret[i / chunksize].begin(), v.begin() + i, v.begin() + endpoint);
 	}
 	return ret;
